Fix S3 upload
- Move the S3 upload to be a service instead of an API route for the
backend. This is because the upload service is used by two other routes,
namely the user's profile image and shopping list cover.
